est.py: use the new planned for drop down field
NEXT-CYCLE has been deprecated. The label will still be supported in the
script, but we also check for the new field that we intend to use.

Also added a dry-run argument to be able to run the script without doing
any updates.

No longer iterating over Initiatives on the ignore list.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>