cloud: add support for cloud server instance
Add a parameter '-c' that implies usage of the cloud server. At this
point this should be seen as a bit of prototype testing. To authenticate
the users to the cloud instance we've also added support for token based
authentication. The updated documentation (config.rst) tells how that
should be done.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>